refactor(backend): standardize Profile constructs names to Model, and Type
